Refactor: Update test dependencies and add new test fakes
This commit introduces several changes to the testing infrastructure:

- Adds new test fakes:
    - `TestBusinessNetworkDataSource`: For testing business-related network operations.
    - `TestInboxNetworkDataSource`: For testing inbox network operations.
    - `TestDatabaseTransactionRunner`: A fake `DatabaseTransactionRunner` for tests.
    - `TestSubscriptionService`: For testing subscription service interactions.
    - `TestBranchDao`: Fake DAO for `BranchEntity`.
    - `TestBusinessDao`: Fake DAO for `BusinessEntity`.
    - `TestLocalSignalDao`: Fake DAO for `LocalSignalEntity`.
    - `TestOutboxCommandDao`: Fake DAO for `OutboxCommandEntity`.
    - `TestSubscriptionDao`: Fake DAO for `SubscriptionEntity`.
    - `TestTaxRateDao`: Fake DAO for `TaxRateEntity`.
- Updates `TestServiceModule` to bind new test service implementations (`TestSubscriptionService`).
- Modifies `TestFirebaseModule` to remove `setPersistenceEnabled(false)` from Firestore settings.
- Updates `BaseTest` to use use cases from the `old` package.
- Makes `failureThrowable` in `BaseTestRepository` protected and adds a `getFailure()` method.
- Updates various test repositories and data files to align with new fakes or remove trailing commas.
- Corrects parameter name in `TestEmployeesBusinessRepository.updateEmployee`.
- Adjusts logic in `TestBusinessRepository` for finding and creating businesses.
- Fixes `creditBalance` update in `TestSubscriptionRepository`.
- Updates `TestSessionRepository` to use `user.uid` for `businessId`.
- Ensures `taxRate.id` is non-null in `TestTaxRepository.addTaxRate`.

+++ core/testing/src/main/kotlin/com/casecode/pos/core/testing/repository/business/TestSubscriptionRepository.kt
@@ -50,7 +50,7 @@ class TestSubscriptionRepository @Inject constructor() : FakeRepository(), Subsc
 
         val currentSub = subscriptions[businessId]
         if (currentSub != null) {
-            val updatedSub = currentSub.copy(creditBalance = currentSub.creditBalance + creditsToAdd)
+            val updatedSub = currentSub.copy(creditBalance = (currentSub.creditBalance + creditsToAdd).toInt())
             subscriptions[businessId] = updatedSub
         }
 
@@ -67,7 +67,7 @@ class TestSubscriptionRepository @Inject constructor() : FakeRepository(), Subsc
         subscriptions[businessId] = subscription
     }
 
-    override  fun clear() {
+    override fun clear() {
         returnSuccess()
         fakeOnboardingConfig = OnboardingConfig(emptyList(), emptyList())
         subscriptions.clear()
